backend/requirements.txt
# -----------------------------
# FastAPI + Server
# -----------------------------
fastapi
uvicorn
python-multipart

# -----------------------------
# Core ML + Data
# -----------------------------
numpy
pandas
scikit-learn
joblib
xgboost
pydantic
PyPDF2
python-docx
# -----------------------------
# LegalBERT
# -----------------------------
transformers

# -----------------------------
# PyTorch (CPU only)
# (need >=2.6.0 to satisfy torch.load security requirement)
# -----------------------------
torch==2.6.0+cpu
torchvision==0.21.0+cpu
torchaudio==2.6.0+cpu
--index-url https://download.pytorch.org/whl/cpu
